--- dlrover/python/unified/remote/call_obj.py
@@ -1,31 +1,0 @@
-# Copyright 2025 The DLRover Authors. All rights reserved.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from dataclasses import dataclass
-
-
-@dataclass
-class RuntimeInfo(object):
-    """
-    Runtime info for workload reporting.
-
-    Attributes:
-        name (str): The unique actor name.
-        create_time (int): The current creation time.
-        hostname (str): The hostname of current actor.
-        host_ip (str): The ip address of current actor.
-    """
-
-    name: str = ""
-    create_time: int = 0
-    hostname: str = ""
-    host_ip: str = ""
